more signed/unsignned fixes
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34201 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ link-grammar/build-disjuncts.c
@@ -531,7 +531,7 @@ void build_sentence_disjuncts(Sentence sent, double cost_cutoff)
 {
 	Disjunct * d;
 	X_node * x;
-	int w;
+	size_t w;
 	for (w = 0; w < sent->length; w++)
 	{
 		d = NULL;
